refactor: renames PureAbility to Ability
BREAKING CHANGE: renames PureAbility to Ability and dops default options. In order, to get the previous behavior of `Ability` class, clients are expected to use `createMongoAbility` and `MongoAbility` type
