+++ packages/casl-ability/src/extra/rulesToFields.ts
@@ -1,11 +1,11 @@
-import { PureAbility } from '../PureAbility';
+import { Ability } from '../Ability';
 import { AnyObject, ExtractSubjectType } from '../types';
 import { setByPath } from '../utils';
 
 /**
  * Extracts rules condition values into an object of default values
  */
-export function rulesToFields<T extends PureAbility<any, AnyObject>>(
+export function rulesToFields<T extends Ability<any, AnyObject>>(
   ability: T,
   action: Parameters<T['rulesFor']>[0],
   subjectType: ExtractSubjectType<Parameters<T['rulesFor']>[1]>,
